requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.4
django-heroku==0.3.1
djangorestframework==3.11.0
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
